refactor(web): inline submissions route page
